Remove Jewler from Rakefile and use Bundler instead
The gem details and dependencies are now specified in the rtex.gemspec.
This makes it possible to use the git repository as a Bundler dependency.
The gem can be build using rake build. Install dependencies using the bundle command.
This adds Bundler as a new dependency!

The VERSION file is not used any more since the gemspec uses
RTeX::Version::STRING instead.